Backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "AI Education Platform Backend"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "asyncpg>=0.29.0",
    "google-generativeai>=0.8.0",
    "gtts>=2.5.0",
    "numpy>=1.26.0",
    "pandas>=2.0.0",
    "psycopg2-binary>=2.9.0",
    "python-dotenv>=1.0.0",
    "scikit-learn>=1.3.0",
    "sqlalchemy>=2.0.0",
]
